Move intrinsic to core::intrinsics::simd
