+++ library/core/src/intrinsics/mod.rs
@@ -3679,62 +3679,3 @@ pub const unsafe fn va_end(ap: &mut VaList<'_>) {
 pub fn return_address() -> *const () {
     core::ptr::null()
 }
-
-/// Returns whether the named target feature is available at the current call site.
-/// This will not emit any runtime detection code; it always turns into `true` or `false` at some
-/// point during compilation.
-///
-/// "Available" in this context means known to be enabled in the calling function.
-/// At a minimum, being enabled globally or in `#[target_features]` of the function containing
-/// this intrinsic is considered available. Depending on codegen options and backend, the features
-/// of the function the intrinsic call is inlined into may also be accounted for. This is intended
-/// to support negligible or zero overhead branching on feature availability, for scenarios where
-/// runtime detection has too much overhead and globally enabled features are not sufficient.
-///
-/// This intrinsic takes a null-terminated feature name passed as a byte array. For a more
-/// ergonomic interface, use the [`target_feature_available_at_call_site!`] macro, which accepts
-/// a string literal.
-///
-/// The LLVM backend implements this intrinsic by lowering to a marker that is resolved by a
-/// post-inlining pass.
-///
-/// # Safety
-///
-/// The return value of this intrinsic is call-site dependent and may vary between calls to the
-/// same function, depending on inlining and codegen. Callers must not rely on the result being
-/// stable across calls or multiple uses of the same function containing the intrinsic.
-///
-/// A return value of `true` guarantees that the feature is enabled at the call site and it is
-/// safe to call other functions that require the feature.
-/// A return value of `false` carries no information and does not indicate whether the feature is
-/// enabled or not.
-#[rustc_nounwind]
-#[unstable(feature = "core_intrinsics", issue = "none")]
-#[rustc_intrinsic]
-pub fn target_feature_available_at_call_site<const FEATURE: [u8; 100]>() -> bool;
-
-/// Returns whether a target feature is enabled at the call site.
-///
-/// This macro is a more convenient interface for [`target_feature_available_at_call_site()`].
-#[unstable(feature = "core_intrinsics", issue = "none")]
-#[allow_internal_unstable(adt_const_params)]
-#[diagnostic::on_unmatched_args(
-    note = "this macro expects a string literal target feature name, like `target_feature_available_at_call_site!(\"avx\")`"
-)]
-#[rustc_macro_transparency = "semiopaque"]
-pub macro target_feature_available_at_call_site($feature:literal) {{
-    ::core::intrinsics::target_feature_available_at_call_site::<
-        {
-            let bytes = $feature.as_bytes();
-
-            assert!(bytes.len() <= 100, "feature string too long");
-
-            // FIXME(const-hack) can't use subslicing yet
-            let mut out = [0u8; 100];
-            let (dst, _) = out.split_at_mut(bytes.len());
-            dst.copy_from_slice(bytes);
-
-            out
-        },
-    >()
-}}
