+++ tests/codegen-llvm/target-feature-available-at-call-site.rs
@@ -4,7 +4,7 @@
 
 #![feature(core_intrinsics)]
 
-use std::intrinsics::target_feature_available_at_call_site;
+use std::intrinsics::simd::target_feature_available_at_call_site;
 
 #[inline]
 pub fn avx_branch_value() -> i32 {
