+++ tests/ui/intrinsics/target-feature-available-at-call-site-const.rs
@@ -1,6 +1,6 @@
 #![feature(core_intrinsics)]
 
-const HAS_FMA: bool = std::intrinsics::target_feature_available_at_call_site!("fma");
+const HAS_FMA: bool = std::intrinsics::simd::target_feature_available_at_call_site!("fma");
 //~^ ERROR cannot call non-const intrinsic `target_feature_available_at_call_site` in constants
 
 fn main() {
