+++ tests/ui/intrinsics/target-feature-available-at-call-site-nonconst.stderr
@@ -5,7 +5,7 @@ LL |     let _ = target_feature_available_at_call_site!(feature);
    |                                                    ^^^^^^^ no rules expected this token in macro call
    |
 note: while trying to match meta-variable `$feature:literal`
-  --> $SRC_DIR/core/src/intrinsics/mod.rs:LL:COL
+  --> $SRC_DIR/core/src/intrinsics/simd/mod.rs:LL:COL
    = note: this macro expects a string literal target feature name, like `target_feature_available_at_call_site!("avx")`
 
 error: aborting due to 1 previous error
